backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
django-filter==25.2
django-unfold==0.90.0
djangorestframework==3.17.1
sqlparse==0.5.5
tzdata==2026.1
gunicorn
whitenoise
dj-database-url
psycopg2-binary
python-dotenv
Pillow
